% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_lasso_cv.R
\name{add_intercept}
\alias{add_intercept}
\title{Adds an intercept to a matrix}
\usage{
add_intercept(mat)
}
\arguments{
\item{mat}{Any matrix}
}
\value{
Matrix with intercept
}
\description{
Adds an intercept to a matrix
}
