% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{dmlmt}
\alias{dmlmt}
\title{Double machine learning for binary and multiple treatments'}
\usage{
dmlmt(x, t, y, pl = TRUE, cs = TRUE, q = 1, cl = NULL, print = FALSE,
  se_rule = NULL, w = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{x}{Matrix of covariates (N x p matrix)}

\item{t}{Vector of treament indicators. Will be ordered from 0 to T-1.}

\item{y}{Vector of outcomes}

\item{pl}{If TRUE Post-Lasso is used to estimate nuisance parameters, if FALSE Lasso}

\item{cs}{If TRUE, common support will be checked}

\item{q}{Quantile used for enforcing common support}

\item{cl}{Vector with cluster variables can be provided}

\item{print}{If TRUE, supporting information is printed}

\item{se_rule}{If not NULL, define, e.g., c(-1,1) to get 1SE and 1SE+ rule}

\item{w}{If TRUE, implied weights are calculated (only if pl=TRUE)}

\item{parallel}{If TRUE, cross-validation of \code{\link{post_lasso_cv}} parallelized}

\item{...}{Pass \code{\link{glmnet}} and \code{\link{post_lasso_cv}} options}
}
\value{
\code{gps_prep} returns a list containing a N x # of treatments matrix
         with generalized propensity scores (\code{p}) and a boolean indicating common
         and a boolean indicating common support (\code{cs}).
}
\description{
This function estimates treatment effects for binary and multiple treatments
using Double Machine Learning.
}
