% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_lasso_cv.R
\name{fitted_values}
\alias{fitted_values}
\title{This function extracts a subset of active variables (nm_act) of the
relavant variables from X'X and X'y to get out-of-sample predictions
for a matrix containing only the active variables
This speeds up the cross-validation for post-Lasso to a large extent}
\usage{
fitted_values(XtX_all, Xty_all, x_pred, nm_act)
}
\arguments{
\item{XtX_all}{Crossproduct of all covariates}

\item{Xty_all}{Crossproduct of covariates and outcome}

\item{x_pred}{Covariates matrix of the PREDICTION sample}

\item{nm_act}{names of active variables}
}
\value{
fit_val fitted values in the prediction sample
}
\description{
This function extracts a subset of active variables (nm_act) of the
relavant variables from X'X and X'y to get out-of-sample predictions
for a matrix containing only the active variables
This speeds up the cross-validation for post-Lasso to a large extent
}
