% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{gps_cs}
\alias{gps_cs}
\title{This function checks common support for a user provided matrix of
(generalized) propensity scores}
\usage{
gps_cs(p_mat, t, q = 1, print = TRUE)
}
\arguments{
\item{p_mat}{N x # of treatments matrix with (generalized) propensity scores}

\item{t}{Matrix of binary treament indicators (N x # of treatments matrix)
each column contains a binary indicator for each tratment}

\item{q}{Quantile used for enforcing common support}

\item{print}{If TRUE, descriptives for p-scores and common support shown}
}
\value{
\code{gps_cs} returns a list containing a N x # of treatments matrix
         with generalized propensity scores (\code{p}) and a boolean indicating common
         and a boolean indicating common support (\code{cs}).
}
\description{
This function checks common support for a user provided matrix of
(generalized) propensity scores
}
