% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{gps_prep}
\alias{gps_prep}
\title{This function prepares the (generalized) propensity scores
and checks the common support after variable selection.}
\usage{
gps_prep(x, t, nm_list, cs = TRUE, q = 1, print = TRUE)
}
\arguments{
\item{x}{Matrix of covariates (N x p matrix)}

\item{t}{Matrix of binary treament indicators (N x # of treatments matrix)
each column contains a binary indicator for each tratment}

\item{nm_list}{List with length = # of treatments containing list with the
selected variable names for each treatment level}

\item{cs}{If TRUE, common support will be checked}

\item{q}{Quantile used for enforcing common support}

\item{print}{If TRUE, descriptives for p-scores and common support shown}
}
\value{
\code{gps_prep} returns a list containing a N x # of treatments matrix
         with generalized propensity scores (\code{p}) and a boolean indicating common
         and a boolean indicating common support (\code{cs}).
}
\description{
This function prepares the (generalized) propensity scores
and checks the common support after variable selection.
}
