% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_lasso_cv.R
\name{norm_w_to_n}
\alias{norm_w_to_n}
\title{Function to normalize weights to N or to N in treated and controls separately}
\usage{
norm_w_to_n(w, d = NULL)
}
\arguments{
\item{w}{vector of weights that should be normalized}

\item{d}{vector of treament indicators}
}
\value{
Normalized weights
}
\description{
Function to normalize weights to N or to N in treated and controls separately
}
