% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{OLS_pred_w}
\alias{OLS_pred_w}
\title{This function calculates the OLS weights for the fitted values of
each observation}
\usage{
OLS_pred_w(x, t, y, cs_i = NULL)
}
\arguments{
\item{x}{Matrix of covariates (N x p matrix)}

\item{t}{Vector of binary treament indicator}

\item{y}{Vector of outcomes}

\item{cs_i}{If not NULL, boolean vector to indicate that observation is on support}
}
\value{
Vector of weights
}
\description{
This function calculates the OLS weights for the fitted values of
each observation
}
