% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{PO_dmlmt}
\alias{PO_dmlmt}
\title{This function calculates the potential outcomes for all treatment levels.}
\usage{
PO_dmlmt(t, y, y_mat, p_mat, cs_i = NULL, cl = NULL, print = TRUE)
}
\arguments{
\item{t}{Matrix of binary treament indicators (N x # of treatments matrix)
each column contains a binary indicator for each tratment}

\item{y}{Vector of outcomes}

\item{y_mat}{N x # of treatments matrix with fitted outcome values}

\item{p_mat}{N x # of treatments matrix with (generalized) propensity scores}

\item{cs_i}{If not NULL, boolean vector to indicate that observation is on support}

\item{cl}{If not NULL, vector with cluster variables}

\item{print}{If TRUE, print results}
}
\value{
\code{PO_dmlmt} returns a list containing the \code{results} and a matrix
         with the individual values of the effcicient score (\code{mu}).
}
\description{
This function calculates the potential outcomes for all treatment levels.
}
