% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{PO_dmlmt_w}
\alias{PO_dmlmt_w}
\title{This function calculates the weights impled by Double Machine learning.}
\usage{
PO_dmlmt_w(x, t, y, p_mat, nm_list, cs_i = NULL, cl = NULL)
}
\arguments{
\item{x}{Matrix of covariates (N x p matrix)}

\item{t}{Matrix of binary treament indicators (N x # of treatments matrix)
each column contains a binary indicator for each tratment}

\item{y}{Vector of outcomes}

\item{p_mat}{N x # of treatments matrix with (generalized) propensity scores}

\item{nm_list}{List with length = # of treatments containing list with the
selected variable names in the outcome regression for each treatment level}

\item{cs_i}{If not NULL, boolean vector to indicate that observation is on support}

\item{cl}{If not NULL, vector with cluster variables}
}
\value{
\code{PO_dmlmt_w} returns a list containing the Double Machine Learning Weights,
        the IPW weights, the OLS weights, and the adjustment weights.
}
\description{
This function calculates the weights impled by Double Machine learning.
}
