% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{PO_dmlmt_w_gen}
\alias{PO_dmlmt_w_gen}
\title{This function calculates the weights impled by Double Machine learning
if outcome weights are already calculated.}
\usage{
PO_dmlmt_w_gen(yw_mat, t, y, p, cs_i = NULL)
}
\arguments{
\item{yw_mat}{Matrix containing weights for each observation (Nt x N matrix)}

\item{t}{Vector of binary treament indicator}

\item{y}{Vector of outcomes}

\item{p}{Vector of propensity scores}

\item{cs_i}{If not NULL, boolean vector to indicate that observation is on support}
}
\value{
Vector of weights
}
\description{
This function calculates the weights impled by Double Machine learning
if outcome weights are already calculated.
}
