% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_lasso_cv.R
\name{post_lasso_cv}
\alias{post_lasso_cv}
\title{This function uses the \code{\link{glmnet}} package to estimate the coefficient paths
and cross-validates Lasso AND Post-Lasso}
\usage{
post_lasso_cv(x, y, w = NULL, kf = 10, family = "gaussian", seed = NULL,
  output = TRUE, d = NULL, parallel = FALSE, graph_file = NULL,
  se_rule = NULL, ...)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes}

\item{w}{vector of weights}

\item{kf}{number of folds in k-fold CV}

\item{family}{Outcome family, currenlty "gaussian" and "binomial" supported}

\item{seed}{CV samples are drawn starting with a specified seed}

\item{output}{If TRUE, output and graphs are printed}

\item{d}{vector with binary treatment indicator, if not NULL,
weights are normalized to sum to N in treated and control also in CV samples}

\item{parallel}{If TRUE, cross-validation parallelized}

\item{graph_file}{Option to define path to save graph}

\item{se_rule}{If not NULL, define, e.g., c(-1,1) to get 1SE and 1SE+ rule}

\item{...}{Pass \code{\link{glmnet}} options}
}
\value{
List with the names of selected variables at cross-validated minima for Lasso and Post-Lasso
}
\description{
This function uses the \code{\link{glmnet}} package to estimate the coefficient paths
and cross-validates Lasso AND Post-Lasso
}
