% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{TE_dmlmt}
\alias{TE_dmlmt}
\title{This function calculates the average treatment effects for all combinations.
Requires input from \code{PO_dmlmt}.}
\usage{
TE_dmlmt(mu, cs_i = NULL, print = TRUE, cl = NULL)
}
\arguments{
\item{mu}{Matrix mu with the individual values of the effcicient score from \code{PO_dmlmt}}

\item{cs_i}{If not NULL, boolean vector to indicate that observation is on support}

\item{print}{If TRUE, print results}

\item{cl}{If not NULL, vector with cluster variables}
}
\value{
\code{TE_dmlmt} returns the results average reatment effects.
}
\description{
This function calculates the average treatment effects for all combinations.
Requires input from \code{PO_dmlmt}.
}
