% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlmt.R
\name{y_prep}
\alias{y_prep}
\title{This function prepares the fitted values of the outcome
after variable selection.}
\usage{
y_prep(x, t, y, nm_list)
}
\arguments{
\item{x}{Matrix of covariates (N x p matrix)}

\item{t}{Matrix of binary treament indicators (N x # of treatments matrix)
each column contains a binary indicator for each tratment}

\item{y}{Vector of outcomes}

\item{nm_list}{List with length = # of treatments containing list with the
selected variable names for each treatment level}
}
\value{
\code{y_prep} returns a N x # of treatments matrix with the fitted values
         for each treatment level.
}
\description{
This function prepares the fitted values of the outcome
after variable selection.
}
