% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_lasso_cv.R
\name{find_Xse_ind}
\alias{find_Xse_ind}
\title{Function finds the positionfor pre-specified SE rules}
\usage{
find_Xse_ind(CV, ind_min, oneSE, factor)
}
\description{
Function finds the positionfor pre-specified SE rules
}
